+++ Astra.Core/src/graphics/renderers/Entity3dRenderer.cpp
@@ -104,7 +104,7 @@ namespace Astra::Graphics
 
 		if (light->GetType() == LightType::Point)
 		{
-			int i = 0;
+			size_t i = 0;
 			for (; i < m_lights.size(); i++)
 			{
 				if (m_lights[i] == light) { break; }
@@ -126,7 +126,7 @@ namespace Astra::Graphics
 		glEnableVertexAttribArray(static_cast<unsigned short>(BufferType::TextureCoords));
 		glEnableVertexAttribArray(static_cast<unsigned short>(BufferType::Normals));
 
-		m_shader->SetUniform1f(NUMBER_OF_ROWS, entity->material->GetRowCount());
+		m_shader->SetUniform1f(NUMBER_OF_ROWS, static_cast<float>(entity->material->GetRowCount()));
 
 		if (entity->material->Transparent)
 		{
